+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 #### Upcoming
 
 *Revisions*
-* Fixed a bug in the nearest config file search algorithm. ([#66](https://github.com/hjwylde/git-fmt/issues/66))
+* Fixed a bug causing the program to hang when not in the root directory. ([#66](https://github.com/hjwylde/git-fmt/issues/66))
 
 #### v0.3.0.3