+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### Upcoming
 
+#### v0.3.1.1
+
 *Revisions*
 * Fixed a bug causing the program to hang when not in the root directory. ([#66](https://github.com/hjwylde/git-fmt/issues/66))
 * Fixed a bug that omitted searching the drive for a config file. ([#66](https://github.com/hjwylde/git-fmt/issues/66))